[refactor] reorganize imports and remove unused helpers
