+++ src/reagentai/agents/main/main_agent.py
@@ -3,11 +3,11 @@ from dataclasses import dataclass
 from aizynthfinder.aizynthfinder import AiZynthFinder
 from pydantic_ai import Tool
 
+from reagentai.tools.image import route_to_image, smiles_to_image
+from src.reagentai.common.aizynthfinder import initialize_aizynthfinder
 from src.reagentai.common.client import LLMClient
 from src.reagentai.constants import AIZYNTHFINDER_CONFIG_PATH
-from src.reagentai.tools.drawing import route_to_image, smiles_to_image
 from src.reagentai.tools.retrosynthesis import (
-    initialize_aizynthfinder,
     perform_retrosynthesis,
 )
 from src.reagentai.tools.smiles import is_valid_smiles
